notebook_requirements.txt
einops==0.4.1
lightning==1.9.0
omegaconf==2.2.3
torchmetrics==0.11.0
torchtyping==0.1.4
tqdm==4.64.0
